pyproject.toml
[project]
name = "git-binance-trader"
version = "0.8.9"
description = "仅模拟盘的币安量化交易系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "jinja2",
  "httpx",
  "pydantic",
  "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
